requirements.txt
transformers>=4.56.1
pre-commit>=4.3.0
openai>=1.108.0
tqdm>=4.67.1
jieba>=0.42.1
requests>=2.28.0
python-dotenv>=1.0.0